+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Fixed setting negative values for memory elements of 4.
 - Memory views are now accessible directly from the "Variable" and "Watch" pane.
 - Data watchpoints can be set directly from the "Watch" pane.
+- Sprites and sprite patterns views visually enhanced.
 
 # 3.7.4
 - PR #193: Fixed bug where pressing 'stop' would sometimes take 5 seconds to disconnect. (mame)